backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
starlette==0.36.3

# ── Rate limiting ─────────────────────────────────────────────────────────────
slowapi==0.1.9

# ── Data & Excel ──────────────────────────────────────────────────────────────
pandas==2.2.1
numpy==1.26.4
openpyxl==3.1.2
xlsxwriter==3.2.0

# ── Security ──────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
passlib==1.7.4               # needed only for sha256_crypt legacy hash verification

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.29
pymysql==1.1.1

# ── Validation ────────────────────────────────────────────────────────────────
pydantic==2.6.4
pydantic[email]==2.6.4       # EmailStr support via email-validator
email-validator==2.1.1

# ── Config ────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── PDF Export ────────────────────────────────────────────────────────────────
reportlab==4.1.0
